src/order-api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
kafka-python==2.0.2
pydantic==2.6.1
